requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.1
pydantic>=2.5.3
email-validator==2.1.1
python-multipart>=0.0.9
httpx>=0.27.0
asyncio>=3.4.3
aiohttp>=3.9.3
sse-starlette>=1.8.2
starlette>=0.46.2
python-logging-loki==0.3.1
motor>=3.3.2
python-dotenv>=1.1.0

